Introduction of conditionally built items
Some items can now be conditionally built, depending on whether their
(findlib) dependencies are satisfied. This is checked using the
findlib library ([Fl_package_base] module).

Note that this feature is not fully operational yet, as the Makefile
will still insist to build all apps and libs. This could be fixed with
the use of an otarget file.