require project name and version in functor arg
We were computing it from the opam file, but actually that is not
recommended by opam especially when you publish the package to
opam-repository. However we do need this info, so now making it
required as part of the plugin’s functor argument. As a side benefit,
we no longer depend on opam-lib.