Changed layout in install directory
each sub-library is now grouped in its own directory in lib/<lib>
instead of being at the root. The META file is changed accordingly and
now has a "directory" variable specifying where to find the
library. Main motivation for that is to avoid having all cmis in the
same directory. This avoid problems with toplevels (and utop in
particular) that could "see" libraries that were not loaded.