reactor a lot
Specifically:
- Factor out some types and functions to new modules Item and Items and
rename many functions therein. Delete Info module.
- Define new Findlib module.
- Delete undeclared_item_check and its uses. This assumed we want to
create a lib and app for every dir or file, respectively, within lib/
and app/ directories. I felt that is too constraining. Clients of this
library should decide that for themselves. It also checked the
converse, that there wasn’t a lib or app declared for which the
corresponding dir or file doesn’t exist. This would indeed be a bug,
but I’m unsure we should be checking for it. A compile error would soon
result anyway, and possibly our check is again too constraining, e.g.
if the dir or file was going to be auto-generated.
- Rename internal Digraph to Graph.
- .ocamlinit: Don’t open Core and Async by default. Now users can
decide to do this by adding it to ocamlinit_prefix.

Tested changes on a couple of projects. It works, but this is still
work in progress. Lots more to do.