delete Items module
Many of the functions regarded searching for a collection of items for
an item with a given name. This was needed prior to
24c00a92cf209a1a2c0b727f42d39f5d58fdd79d because the dependencies of an
item were stored by their name only. However, since that commit, the
dependencies are themselves values of type item. Thus, searching for
items by name is no longer needed in all the cases we were using it for.

A handful of functions are still needed, and in particular we do still
want a `private Item.t list`, on which we can assume some invariants.
This type and a few other operations were moved from Items to Item.