build pack deps dynamically
Now setting the deps of the rule to be ml files instead of
corresponding cmo files. This assures just that the ml files get
created. Thus within the build action, we can now be sure all the ml
files exist and call ocamldep. Finally, we dynamically build the cmo
files. This goes further towards resolving #14.