force cmo/cmx compilation order
Ocamlbuild’s builder takes a list of targets to compile, and tries as
much as possible to build the targets in parallel. Even with -j not
set, one cannot assume that the items will be compiled in the order
given (I experienced this in a larger project). We now manually force
the build to occur in order, at the expense of not allowing parallel
builds of these items even if possible.