Project: make C libs a deb of byte and native libs
In 676ff9c83d29f5d37915568446617739b0a8d003, we separated out the build
of C, byte, and native libs. As a result, there was nothing actually
triggering the build of the C libs. Since the byte and native libs have
to refer or link in the C libs, it seems sensible to make the C libs a
dependency of them.