OCaml: rename capital flags
Unix flags starting with a capital letter can’t be mapped directly to
an OCaml labeled argument, which can’t start with a capital letter.
Were resolving this be using an initial underscore, but that is a bad
choice because OCaml treats variables so named specially. Now handling
this differently, and added documentation to explain.