OCaml: fix redundant command line flags
Several eliom tools take arguments that overlap with the tool the eliom
tool is a wrapper around. Thus, the flags end up being defined twice.
That has either been handled correctly by eliom or it is unclear what
happens in these cases if the flag is given twice. Now, our code at
least passes all the flags, instead of ignoring some as before. This
will hopefully expose bugs and force us and or eliom to be fixed.