Project.path_of_packed_lib: rename and change arg
Rename to path_of_pack. Saying “lib” is misleading because a pack leads
to a module, not a lib. Also drop the pack_name arg since it is
inferable from the lib arg. This does mean the function can now raise
an exception whereas it couldn’t before, but this still seems cleaner.
Finally, expose the function and document it.