Project: improve plugin configuration
Now providing build_plugin option instead of shared. For one thing,
were failing to use the `shared` option in any way. It was a misleading
name. This isn’t really a compiler flag that is up to the user, like
safe_string. Rather, the user can either select to build a cmxs plugin
or not, and if they do then internally we need to use -shared option.