don't require libs
It is now okay to have projects that build only an app, and no libs. In
this case, we make sure that META file support is disabled. That means
we don’t register a rule to build a META file, don’t include it in the
.install file, nor insert a rule to build it in the generated
project.mk file.

Return type of meta_file is now an option, representing the fact that
you won’t get an answer if the number of libs is 0. Also provided a
build_meta_file for convenience internally.