fix -inline option to ocamlopt
Now using string type instead of int. This makes it consistent with all
the other inline* flags. String is better than int because the
arguments can be more complex than just an int.

Tools.ocamlopt_args_specs: There was a bug. I was creating a command
with the -inline arg given twice.