Project.build_static_file: allow lazy building of content
We now take a function to build the content, instead of the content
itself. Thus, the content is now created only if the registered rule is
actually run. This can improve performance if content creation takes a
non-negligible amount of time. Rules always get registered but not
necessarily run if you don’t ask for that target to be built.