rename `findlib_pkg` field to `install`
The `findlib_pkg` field of a lib is now called `install`. You can now
either specify `No to mean the library should not be installed or
`Findlib x to indicated it should be installed according to the
conventions of findlib with package name x.

Renamed type pkg to findlib_pkg to emphasize this is a findlib package
as opposed to an opam package.